Show contract errors unless it is the one signalling to start verification
The message `Verifiable Credential: Transfers of this amount require
validateAndTransfer" is only shown when a signer tries to call
`transfer` over the threshold. Therefore, we will skip showing an error
in that scenario as we handle it. Otherwise, we want to continue showing
errors for other contract errors beginning with the string "Verifiable
Credential" such as error messages when the verification is expired,
etc.